fix : 칵테일 삭제 api 예외처리

+++ src/main/java/salute/oneshot/domain/cocktail/service/CocktailService.java
@@ -1,5 +1,6 @@
 package salute.oneshot.domain.cocktail.service;
 
+import java.nio.file.AccessDeniedException;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -20,6 +21,8 @@ import salute.oneshot.domain.ingredient.entity.Ingredient;
 import salute.oneshot.domain.ingredient.repository.IngredientRepository;
 import salute.oneshot.domain.user.entity.User;
 import salute.oneshot.domain.user.repository.UserRepository;
+import salute.oneshot.global.exception.NotFoundException;
+import salute.oneshot.global.exception.UnauthorizedException;
 
 @Service
 @RequiredArgsConstructor
@@ -41,10 +44,11 @@ public class CocktailService {
         cocktailRepository.save(cocktail);
 
         List<CocktailIngredient> ingredientList = sDto.getIngredientList().stream()
-            .map( req-> {
-                    Ingredient ingredient = ingredientRepository.getReferenceById(req.getIngredientId());
-                    return CocktailIngredient.of(cocktail, ingredient , req.getVolume());
-                }).toList();
+            .map(req -> {
+                Ingredient ingredient = ingredientRepository.getReferenceById(
+                    req.getIngredientId());
+                return CocktailIngredient.of(cocktail, ingredient, req.getVolume());
+            }).toList();
 
         cocktailIngredientRepository.saveAll(ingredientList);
 
@@ -53,6 +57,10 @@ public class CocktailService {
 
     @Transactional
     public void deleteCocktail(DeleteCocktailSDto sDto) {
+
+        if (!cocktailRepository.existsByIdAndUserId(sDto.getCocktailId(), sDto.getUserId())) {
+            throw new UnauthorizedException(ErrorCode.FORBIDDEN_ACCESS);
+        }
         cocktailRepository.deleteById(sDto.getCocktailId());
     }
 }
